+++ packages/ui/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.31.17](https://github.com/ttoss/ttoss/compare/@ttoss/ui@1.31.16...@ttoss/ui@1.31.17) (2023-04-13)
+
+**Note:** Version bump only for package @ttoss/ui
+
 ## [1.31.16](https://github.com/ttoss/ttoss/compare/@ttoss/ui@1.31.15...@ttoss/ui@1.31.16) (2023-04-12)
 
 ### Bug Fixes